+++ packages/eas-cli/src/submit/context.ts
@@ -33,6 +33,7 @@ export interface SubmissionContext<T extends Platform> {
   analytics: Analytics;
   vcsClient: Client;
   applicationIdentifierOverride?: string;
+  whatToTest?: string;
   specifiedProfile?: string;
 }
 
@@ -60,6 +61,7 @@ export async function createSubmissionContextAsync<T extends Platform>(params: {
   exp: ExpoConfig;
   projectId: string;
   vcsClient: Client;
+  whatToTest?: string;
   specifiedProfile?: string;
 }): Promise<SubmissionContext<T>> {
   const {