+++ packages/eas-cli/src/submit/ios/IosSubmitter.ts
@@ -125,12 +125,13 @@ export default class IosSubmitter extends BaseSubmitter<
   ): IosSubmissionConfigInput {
     const { appSpecificPassword, ascApiKeyResult } = credentials;
     const { appleIdUsername, ascAppIdentifier } = options;
-    const { isVerboseFastlaneEnabled, groups } = this.ctx;
+    const { isVerboseFastlaneEnabled, groups, whatToTest } = this.ctx;
     return {
       ascAppIdentifier,
       appleIdUsername,
       isVerboseFastlaneEnabled,
       groups,
+      changelog: whatToTest,
       ...(appSpecificPassword ? this.formatAppSpecificPassword(appSpecificPassword) : null),
       ...(ascApiKeyResult?.result ? this.formatAscApiKeyResult(ascApiKeyResult.result) : null),
     };